+++ packages/react-native-sdk/src/components/Participant/FloatingParticipantView/index.tsx
@@ -147,7 +147,7 @@ export const FloatingParticipantView = ({
             {ParticipantView && (
               <ParticipantView
                 participant={participant}
-                videoMode={'video'}
+                trackType="videoTrack"
                 style={[
                   styles.participantViewContainer,
                   style,