+++ packages/react-sdk/src/core/components/ParticipantView/ParticipantView.tsx
@@ -12,8 +12,8 @@ import {
   SfuModels,
   StreamVideoLocalParticipant,
   StreamVideoParticipant,
+  VideoTrackType,
 } from '@stream-io/video-client';
-import { useCallStateHooks } from '@stream-io/video-react-bindings';
 
 import { Audio } from '../Audio';
 import { Video, VideoProps } from '../Video';
@@ -22,7 +22,7 @@ import { DefaultParticipantViewUI } from './DefaultParticipantViewUI';
 import { applyElementToRef, isComponentType } from '../../../utilities';
 
 export type ParticipantViewContextValue = Required<
-  Pick<ParticipantViewProps, 'participant' | 'videoMode'>
+  Pick<ParticipantViewProps, 'participant' | 'trackType'>
 > & {
   participantViewElement: HTMLDivElement | null;
   videoElement: HTMLVideoElement | null;
@@ -52,7 +52,7 @@ export type ParticipantViewProps = {
   /**
    * The kind of video stream to play for the given participant. The default value is `video`. You can use `none` if you're building an audio-only call.
    */
-  videoMode?: 'video' | 'screen' | 'none';
+  trackType?: VideoTrackType | 'none';
 
   /**
    * This prop is only useful for advanced use-cases (for example building your own paginated layout). When set to `true` it will mute the give participant's audio stream on the client side. The local participant is always muted.
@@ -77,7 +77,7 @@ export const ParticipantView = forwardRef<HTMLDivElement, ParticipantViewProps>(
   (
     {
       participant,
-      videoMode = 'video',
+      trackType = 'videoTrack',
       muteAudio,
       refs: { setVideoElement, setVideoPlaceholderElement } = {},
       className,
@@ -87,15 +87,12 @@ export const ParticipantView = forwardRef<HTMLDivElement, ParticipantViewProps>(
     ref,
   ) => {
     const {
-      audioStream,
       isLocalParticipant,
       isSpeaking,
       isDominantSpeaker,
       publishedTracks,
       sessionId,
     } = participant;
-    const { useLocalParticipant } = useCallStateHooks();
-    const localParticipant = useLocalParticipant();
 
     const hasAudio = publishedTracks.includes(SfuModels.TrackType.AUDIO);
     const hasVideo = publishedTracks.includes(SfuModels.TrackType.VIDEO);
@@ -114,6 +111,7 @@ export const ParticipantView = forwardRef<HTMLDivElement, ParticipantViewProps>(
     useTrackElementVisibility({
       sessionId,
       trackedElement,
+      trackType,
     });
 
     const participantViewContextValue = useMemo(
@@ -122,14 +120,14 @@ export const ParticipantView = forwardRef<HTMLDivElement, ParticipantViewProps>(
         participantViewElement: trackedElement,
         videoElement: contextVideoElement,
         videoPlaceholderElement: contextVideoPlaceholderElement,
-        videoMode,
+        trackType,
       }),
       [
         contextVideoElement,
         contextVideoPlaceholderElement,
         participant,
         trackedElement,
-        videoMode,
+        trackType,
       ],
     );
 
@@ -165,15 +163,12 @@ export const ParticipantView = forwardRef<HTMLDivElement, ParticipantViewProps>(
         <ParticipantViewContext.Provider value={participantViewContextValue}>
           {/* mute the local participant, as we don't want to hear ourselves */}
           {!isLocalParticipant && !muteAudio && (
-            <Audio
-              sinkId={localParticipant?.audioOutputDeviceId}
-              audioStream={audioStream}
-            />
+            <Audio participant={participant} />
           )}
           <Video
             VideoPlaceholder={VideoPlaceholder}
             participant={participant}
-            kind={videoMode}
+            trackType={trackType}
             refs={videoRefs}
             autoPlay
           />